+++ Core/PeaksFinder.cpp
@@ -1,7 +1,7 @@
 #include "PeaksFinder.h"
 
 Utils::FixedSizePQ<Core::Peak, Consts::Fingerprint::NPeaks>
-Core::findPeaks(const Math::FFTWindow &fftWindow, const size_t &window, const int &bandStart, const int &bandEnd) {
+Core::findPeaks(const Math::FFTWindow &fftWindow, const int64_t &window, const int &bandStart, const int &bandEnd) {
     Utils::FixedSizePQ<Peak, Consts::Fingerprint::NPeaks> peaks;
     float magCurrent, freqCurrent;
     int indexLeft, indexRight, j;
@@ -18,7 +18,7 @@ Core::findPeaks(const Math::FFTWindow &fftWindow, const size_t &window, const in
         magCurrent = fftWindow.getMagnitudes()[i];
 
         /*
-         * Extract respectively 5 element after and before i.
+         * Extract respectively 5 element after and before 'i'.
          * Are there 5 element before and after actually?
          */
         indexLeft = i - Consts::Fingerprint::PeakRange >= bandStart ? i - Consts::Fingerprint::PeakRange : bandStart;
@@ -33,7 +33,7 @@ Core::findPeaks(const Math::FFTWindow &fftWindow, const size_t &window, const in
         //same as above
         for (j = i + 1; j <= indexRight && (ok = fftWindow.getMagnitudes()[j] < magCurrent); j++);
         if (!ok) {
-            //if @ j there's a loudest peak, we can skip all the peaks between i and j-1
+            //if @ j there's a loudest peak, we can skip all the peaks between 'i' and 'j-1'
             i = j - 1; //for will increment it
             continue;
         }
